[AS7-6263] Allow expressions for management core service
* allow expressions for http-interfaces's console-enabled attribute

* renamed HostModelTests to HostModelTestCase
* the test was not picked by the test suite. Once renamed, it was
  failing because the domain-controller was not sanitized for the XML
  output
* add test with host.xml variant with expressions
